+++ spring-kotlin-coroutine/src/main/java/org/springframework/kotlin/experimental/coroutine/util/CoroutineUtils.java
@@ -18,15 +18,16 @@ package org.springframework.kotlin.experimental.coroutine.util;
 
 import kotlin.coroutines.experimental.Continuation;
 import kotlin.coroutines.experimental.CoroutineContext;
-import kotlin.jvm.functions.Function1;
+import kotlin.jvm.functions.Function2;
 import kotlinx.coroutines.experimental.BuildersKt;
+import kotlinx.coroutines.experimental.CoroutineScope;
 import kotlinx.coroutines.experimental.CoroutineStart;
 
 public abstract class CoroutineUtils {
     private CoroutineUtils() {
     }
 
-    static public <T> Object runCoroutine(final CoroutineContext context, final Function1<Continuation<? super T>, ? super T> fun, final Continuation<? super T> continuation) {
+    static public <T> Object runCoroutine(final CoroutineContext context, final Function2<CoroutineScope, Continuation<? super T>, ? super T> fun, final Continuation<? super T> continuation) {
         return BuildersKt.withContext(context, CoroutineStart.DEFAULT, fun, continuation);
     }
 }