+++ spring-kotlin-coroutine/src/test/groovy/org/springframework/kotlin/experimental/coroutine/scheduler/CoroutineSchedulerSpec.groovy
@@ -46,7 +46,7 @@ class CoroutineSchedulerSpec extends Specification {
 
         where:
         configuration << [FixedDelayConfiguration, FixedDelayStringConfiguration,
-                          FixedDelayOnCommonPoolConfiguration, FixedDelayWithSuspensionPointConfiguration]
+                          FixedDelayOnDefaultDispatcherConfiguration, FixedDelayWithSuspensionPointConfiguration]
     }
 
     @Unroll